Arreglo prioridad de operadores, juego calcular
Para el nivel 2 ya reconoce la prioridad, falta probar y mejorar la
prioridad del nivel 1. Falla en el temporizador.
Mejore para mostrar los aciertos y los puntos que lleva.

+++ BrainGame/app/src/main/java/com/sdm/uniovi/braingame/juegos/calcular/logica/Constante.java
@@ -9,7 +9,7 @@ public class Constante implements Expresion {
     }
 
     @Override
-    public double valor() {
+    public int valor() {
         return valor;
     }
