+++ BrainGame/app/src/main/java/com/sdm/uniovi/braingame/juegos/calcular/logica/Expresion.java
@@ -4,6 +4,6 @@ public interface Expresion {
 
 
 
-     double valor();
+     int valor();
     String mostrar();
 }
